add option to include changelog

+++ debian/copy_from_pkg.sh
@@ -3,6 +3,9 @@
 # KA Lite sources are really big, this script takes an original KA Lite
 # debian source and creates one without all the fuzz.
 
+# If you did changes in a package and want the changelog included:
+# ./copy_from_package /path/to/ka-lite-source-0.16~b1 -c
+
 set -e
 
 if
@@ -24,6 +27,13 @@ then
 	test_dir=test/ka-lite-test
 else
 	test_dir=$1
+fi
+
+if [ "$2" == "-c" ]
+then
+	also_changelog=1
+else
+	also_changelog=0
 fi
 
 dest_dir=`realpath .`
@@ -55,7 +65,10 @@ case "$choice" in
 		# The below has to run with respect to relative paths, hence the cd and the cp --parents
 		cd $test_dir
 		find debian -mindepth 1 -not -name '*.zip' -exec cp -rp \{\} --parents $dest_dir \;
-		rm $dest_dir/debian/changelog
+		if [ $also_changelog -eq 0 ]
+		then
+			rm $dest_dir/debian/changelog
+		fi
 		git reset -- $dest_dir/debian/changelog
 		git checkout $dest_dir/debian/changelog
 ;;
